move stack code into its own file
